+++ src/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["main", "manager", "executor", "functions", "prologue", "graphical_object"]