+++ src/model_config/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["graphic_object", "model", "types"]