+++ src/model_config/types.py
@@ -1,57 +1,14 @@
+"""
+The 3 basic node types and then the MFHTS Implementation #1
+"""
+import math
+
+
 OCT = "0o"  # Octal: 0-7
 HEX = "0x"  # Hexadecimal: 0-F
 BIN = "0b"  # Binary: 0, 1
 INT = "0i"  # Integer: 0-9
 FLT = "0f"  # Floating-Point Number: 0-9 w/ decimal
-
-# class Input and Output are semi-dummy class to differntiate between the nodes and types.
-
-class Input:
-    def __init__(self, typ):
-        """
-        :param typ: Inherited from <Any> class
-        """
-        self.typ = typ
-
-    # to read by the people
-    def __str__(self):
-        return "input type"
-
-    # to be read by the executor
-    def __repr__(self):
-        return "i"
-
-class Output:
-    def __init__(self, typ):
-        """
-        :param typ: Inherited from <Any> class
-        """
-        self.typ = typ
-
-    # to read by the people
-    def __str__(self):
-        return "output type"
-
-    # to be read by the executor
-    def __repr__(self):
-        return "o"
-
-class Constant:
-    def __init__(self, exe):
-        """
-        :param exe: A class that executes the custom fields, aka constants
-        """
-        self.exe = exe
-
-    # to read by the people
-    def __str__(self):
-        return "constant type"
-
-    # to be read by the executor
-    def __repr__(self):
-        return "c"
-
-
 
 
 # any type
@@ -63,47 +20,84 @@ class Any:
         return str(self.__class__.__name__)+": "+str(self.DATA)
 
 
-# any number from float to binary
+class Void(Any):
+    def __init__(self):
+        super().__init__(None)
+
+    def __str__(self):
+        return "Void"
+
+
 class Number(Any):
-    def __init__(self, dt=None):
-        prefix = dt[:2]
-        number = dt[2:]
-        if type(dt) in [int, float]:
-            self.DATA = dt
-        elif prefix in [OCT, HEX, BIN, INT, FLT]:
-            if prefix == "0o":
-                for c in number:
-                    pass
-            self.DATA = dt
-        raise ValueError(f"Value {dt} is not number")
+    def __init__(self, n):
+        if type(n) in [int, float]:
+            super().__init__(n)
+        else:
+            raise ValueError(f"Error: The class <Number> received an invalid number: {n}")
 
     def __str__(self):
         return str(self.DATA)
 
+    def __add__(self, other): return self.DATA+other.DATA
 
-# any type that consists of group of another type
+    def __sub__(self, other): return self.DATA-other.DATA
+
+    def __mul__(self, other): return self.DATA-other.DATA
+
+    def __truediv__(self, other):
+        if other.DATA != 0: return self.DATA/other.DATA
+        else: return 0
+
+    def __floordiv__(self, other): return self.DATA//other.DATA
+
+    def ceil(self, other): return math.ceil(self.DATA/other.DATA)
+
+    def __mod__(self, other): return self.DATA%other.DATA
+
+    def __pow__(self, power, modulo=None): return self.DATA**power.DATA
+
+    def __lt__(self, other): return self.DATA < other.DATA
+
+    def __le__(self, other): return self.DATA <= other.DATA
+
+    def __eq__(self, other): return self.DATA == other.DATA
+
+    def __ne__(self, other): return self.DATA != other.DATA
+
+    def __ge__(self, other): return self.DATA >= other.DATA
+
+
 class Collections(Any):
-    def __init__(self, dt=None):
-        prefix = dt[:2]
-        number = dt[2:]
-        if type(dt) in [int, float]:
-            self.DATA = dt
-        elif prefix in ["0o", "0x", "0b"]:
-            if prefix == "0o":
-                pass
-            self.DATA = dt
-        raise ValueError(f"Value {dt} is not number")
+    def __init__(self, *args):
+        super().__init__(args)
 
-    def __str__(self):
-        return str(self.DATA)
+    def __len__(self): return len(self.DATA)
 
 
-class Integer(Number):
-    def __init__(self, dt=None):
-        if type(dt) not in [Integer]:
-            pass
-        self.DATA = dt
 
-    def __str__(self):
-        pass
+def numberTest():
+    a = Number(45)
+    b = Number(5)
+    c = Number(3.9)
+    print(a+b)
+    print(a-b)
+    print(a*b)
+    print(a/b)
+    print(a/Number(0))
+    print(a/c)
+    print(a//c)
+    print(a.ceil(c))
+    print(a%b)
+    print(a**b)
+    print(a>b)
+    print(a>=b)
+    print(a==b)
+    print(a<=b)
+    print(a<b)
+    print([0, 0, 0, 1, 1] > [0, 0])
 
+    a += Number(3)
+
+a = Collections(1, 3, 4, 3, 6, 3)
+print(len(a))
+print()