+++ src/test1/awesome_one_liners.py
@@ -1,0 +1,5 @@
+
+# to remove a set of character from a string
+s = "Random Text"
+dat = s.translate({ord(i): None for i in 'characters to be removed'})
+print(dat)