+++ travis/test.py
@@ -7,3 +7,9 @@ print(sys.version[:5])
 import sklearn
 import pygame
 
+test_variable_for_fstring = 17
+print(f"{test_variable_for_fstring} and 31")
+
+############################################
+
+